+++ src/KReg.cpp
@@ -22,7 +22,7 @@
 #define DEFAULT_OVECTOR	30
 KRegSubString::~KRegSubString()
 {
-	if(subStrings){
+	if (subStrings) {
 		for (int i = 0; i < count; i++) {
 			if (subStrings[i]) {
 				xfree(subStrings[i]);
@@ -32,14 +32,14 @@ KRegSubString::~KRegSubString()
 	}
 }
 KReg::KReg() {
-	model=NULL;
-	c_model=NULL;
-	pe=NULL;
-//	int match_limit=1;
-//	pcre_config(PCRE_CONFIG_MATCH_LIMIT,&match_limit);
+	model = NULL;
+	c_model = NULL;
+	pe = NULL;
+	//	int match_limit=1;
+	//	pcre_config(PCRE_CONFIG_MATCH_LIMIT,&match_limit);
 }
 KReg::~KReg() {
-	if (model!=NULL) {
+	if (model != NULL) {
 		xfree(model);
 	}
 	if (c_model) {
@@ -49,28 +49,28 @@ KReg::~KReg() {
 		freeStudy(pe);
 	}
 }
-KRegSubString *KReg::matchSubString(const char *str,int str_len,int flag,int *ovector,int ovector_size)
+KRegSubString* KReg::matchSubString(const char* str, int str_len, int flag, int* ovector, int ovector_size)
 {
 	int matched = match(str, str_len, 0, ovector, ovector_size);
 	if (matched < 1) {
 		return NULL;
 	}
-	return makeSubString(str,ovector,ovector_size,matched);
+	return makeSubString(str, ovector, ovector_size, matched);
 }
-KRegSubString *KReg::matchSubString(const char *str,int str_len,int flag)
+KRegSubString* KReg::matchSubString(const char* str, int str_len, int flag)
 {
 	int ovector[OVECTOR_SIZE];
 	int matched = match(str, str_len, 0, ovector, OVECTOR_SIZE);
 	if (matched < 1) {
 		return NULL;
 	}
-	return makeSubString(str,ovector,OVECTOR_SIZE,matched);
+	return makeSubString(str, ovector, OVECTOR_SIZE, matched);
 }
 bool KReg::isPartialModel() {
 #ifdef PCRE_INFO_OKPARTIAL
-	int okPartial=0;
-	if (pcre_fullinfo(c_model, pe, PCRE_INFO_OKPARTIAL, &okPartial)==0) {
-		if (okPartial==1) {
+	int okPartial = 0;
+	if (pcre_fullinfo(c_model, pe, PCRE_INFO_OKPARTIAL, &okPartial) == 0) {
+		if (okPartial == 1) {
 			return true;
 		}
 	}
@@ -81,24 +81,24 @@ bool KReg::isPartialModel() {
 
 }
 
-bool KReg::setModel(const char *model_str, int flag,int match_limit) {
-	if (model_str==NULL||*model_str=='\0'){
+bool KReg::setModel(const char* model_str, int flag, int match_limit) {
+	if (model_str == NULL || *model_str == '\0') {
 		return false;
 	}
-	if (model!=NULL) {
+	if (model != NULL) {
 		xfree(model);
-		model=NULL;
+		model = NULL;
 	}
 	if (c_model) {
 		pcre_free(c_model);
-		c_model=NULL;
+		c_model = NULL;
 	}
-	model=xstrdup(model_str);
-	const char *error = NULL;
+	model = xstrdup(model_str);
+	const char* error = NULL;
 	int erroffset;
-	c_model=pcre_compile(model_str, flag, &error, &erroffset, NULL);
+	c_model = pcre_compile(model_str, flag, &error, &erroffset, NULL);
 	if (c_model) {
-		const char * error;
+		const char* error;
 		if (pe) {
 			freeStudy(pe);
 		}
@@ -109,71 +109,71 @@ bool KReg::setModel(const char *model_str, int flag,int match_limit) {
 			0
 #endif
 			,
-		&error); /* set to NULL or points to a message */
-		if(pe && match_limit>0){
-			pe->flags|=PCRE_EXTRA_MATCH_LIMIT;
+			&error); /* set to NULL or points to a message */
+		if (pe && match_limit > 0) {
+			pe->flags |= PCRE_EXTRA_MATCH_LIMIT;
 			pe->match_limit = match_limit;
 		}
 		return true;
 	}
-	klog(KLOG_ERR,"cann't compile regex [%s] pos=%d,error=[%s]\n",model,erroffset,(error?error:""));
+	klog(KLOG_ERR, "cann't compile regex [%s] pos=%d,error=[%s]\n", model, erroffset, (error ? error : ""));
 	return false;
 }
-const char *KReg::getModel() {
+const char* KReg::getModel() {
 	if (model)
 		return model;
 	return "";
 }
-int KReg::matchPartial(const char *str, int str_len, int flag, int *workspace,
-		int wscount) {
+int KReg::matchPartial(const char* str, int str_len, int flag, int* workspace,
+	int wscount) {
 	int ovector[DEFAULT_OVECTOR];
 	return matchPartial(str, str_len, flag, ovector, DEFAULT_OVECTOR,
-			workspace, wscount);
+		workspace, wscount);
 }
-int KReg::matchNext(const char *str, int str_len, int flag, int *workspace,
-		int wscount) {
+int KReg::matchNext(const char* str, int str_len, int flag, int* workspace,
+	int wscount) {
 	int ovector[DEFAULT_OVECTOR];
 	return matchNext(str, str_len, flag, ovector, DEFAULT_OVECTOR, workspace,
-			wscount);
+		wscount);
 }
-int KReg::matchPartial(const char *str, int str_len, int flag, int *ovector,
-		int ovector_size, int *workspace, int wscount) {
+int KReg::matchPartial(const char* str, int str_len, int flag, int* ovector,
+	int ovector_size, int* workspace, int wscount) {
 	//	return -1;
-	#ifdef PCRE_DFA_SHORTEST
-	return pcre_dfa_exec(c_model, pe, str, str_len, 0, 
-	PCRE_PARTIAL|flag, ovector, ovector_size, workspace, wscount);
-	#else
+#ifdef PCRE_DFA_SHORTEST
+	return pcre_dfa_exec(c_model, pe, str, str_len, 0,
+		PCRE_PARTIAL | flag, ovector, ovector_size, workspace, wscount);
+#else
 	return -1;
-	#endif
+#endif
 }
-int KReg::matchNext(const char *str, int str_len, int flag, int *ovector,
-		int ovector_size, int *workspace, int wscount) {
-	        #ifdef PCRE_DFA_SHORTEST
+int KReg::matchNext(const char* str, int str_len, int flag, int* ovector,
+	int ovector_size, int* workspace, int wscount) {
+#ifdef PCRE_DFA_SHORTEST
 
-	return pcre_dfa_exec(c_model, pe, str, str_len, 0, 
-	PCRE_DFA_RESTART|flag, ovector, ovector_size, workspace, wscount);
-	#else
+	return pcre_dfa_exec(c_model, pe, str, str_len, 0,
+		PCRE_DFA_RESTART | flag, ovector, ovector_size, workspace, wscount);
+#else
 	return -1;
-	#endif
+#endif
 }
 
-int KReg::match(const char *str, int str_len, int flag) {
+int KReg::match(const char* str, int str_len, int flag) {
 	int ovector[DEFAULT_OVECTOR];
 	return match(str, str_len, flag, ovector, DEFAULT_OVECTOR);
 }
 
-int KReg::match(const char *str, int str_len, int flag, int *ovector,
-		int ovector_size) {
-	if (str_len==-1) {
-		str_len=strlen(str);
+int KReg::match(const char* str, int str_len, int flag, int* ovector,
+	int ovector_size) {
+	if (str_len == -1) {
+		str_len = (int)strlen(str);
 	}
 	return pcre_exec(c_model, /* result of pcre_compile() */
-	pe, /* we didn't study the pattern */
-	str, /* the subject string */
-	str_len, /* the length of the subject string */
-	0, /* start at offset 0 in the subject */
-	flag, /* default options */
-	ovector, /* vector of integers for substring information */
-	ovector_size); /* number of elements in the vector (NOT size in bytes) */
+		pe, /* we didn't study the pattern */
+		str, /* the subject string */
+		str_len, /* the length of the subject string */
+		0, /* start at offset 0 in the subject */
+		flag, /* default options */
+		ovector, /* vector of integers for substring information */
+		ovector_size); /* number of elements in the vector (NOT size in bytes) */
 }
